requirements.txt
pyexiftool
click
